Fix the curve editor and leaf materials
- Assigning leaves materials caused a segmentation fault because geometry
  segments were not passed by reference and thus were never updated.
- The curve editor was not updated on undo operations outside of the widget.
- Undo and redo operations did not work properly in the curve editor because
  commands operated on the modified spline instead of the original spline.

+++ plant_generator/mesh.h
@@ -52,7 +52,7 @@ namespace pg {
 		void addLeaves(Stem *);
 		void addLeaf(Leaf *leaf, Stem *stem);
 		void addTriangle(int, int, int);
-		int selectBuffer(int);
+		int selectBuffer(long);
 		void initBuffer();
 		void updateSegments();
